+++ src/main/(sockets)/chats/dto/chat.event.ts
@@ -1,0 +1,7 @@
+export type ChatEvents = {
+  "chat:message": {
+    roomId: string;
+    senderId: string;
+    content: string;
+  };
+};