+++ src/main/(started)/auth/auth.repository.ts
@@ -12,10 +12,7 @@ export class AuthRepository {
   async validateUserWithPassword(email: string, password: string) {
     const user = await this.userRepository.findByEmail(email);
     if (!user) throw new NotFoundException("User not found with that email");
-    const compare = await this.utilityService.compare(
-      user.passwordHash!,
-      password,
-    );
+    const compare = await this.utilityService.compare(user.password!, password);
     if (!compare) throw new NotFoundException("Invalid credentials");
     return user;
   }